# cedar-valuerecommender-server-application/src/main/resources/assets/swagger-api/swagger.yaml
openapi: 3.0.1
info:
  description: CEDAR Value Recommender Server API
  title: CEDAR Value Recommender Server API
  version: 2.9.11
paths:
  /:
    get:
      operationId: showInfo
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The service name and links to its API documentation
      summary: Describe this CEDAR service
      tags:
      - Service
  /command/can-generate-recommendations:
    post:
      description: Check whether the value recommender can generate recommendations
        for a template (or for any template if none is provided).
      operationId: areRecommendationsEnabled
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
        "503":
          description: OpenSearch unavailable
      security:
      - api_key: []
      summary: Check whether recommendations can be generated
      tags:
      - Command
  /command/generate-rules/status:
    get:
      description: Get status information about the rule-generation process for all
        templates.
      operationId: getRulesGenerationStatusAll
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - api_key: []
      summary: Get rule-generation status for all templates
      tags:
      - Command
  /command/generate-rules/status/{template_id}:
    get:
      description: Get status information about the rule-generation process for a
        template.
      operationId: getRulesGenerationStatus
      parameters:
      - description: Template identifier.
        in: path
        name: template_id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - api_key: []
      summary: Get rule-generation status for a template
      tags:
      - Command
  /command/generate-rules/{template_id}:
    post:
      description: Generate the mining rules that the value recommender uses to produce
        recommendations for a template.
      operationId: generateRules
      parameters:
      - description: Template identifier.
        in: path
        name: template_id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
      security:
      - api_key: []
      summary: Generate rules for a template
      tags:
      - Command
  /command/recommend:
    post:
      description: Get metadata recommendations for a target field.
      operationId: recommendValues
      requestBody:
        content:
          '*/*':
            schema:
              $ref: "#/components/schemas/RecommendationInput"
        description: The recommendation input
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "500":
          description: Internal server error
        "503":
          description: OpenSearch unavailable
      security:
      - api_key: []
      summary: Get recommendation
      tags:
      - Command
  /healthcheck:
    get:
      operationId: healthCheck
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: Every health check passed
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                type: object
          description: At least one health check failed
      security:
      - api_key: []
      summary: Run this service's health checks
      tags:
      - Service diagnostics
  /insight/full:
    get:
      operationId: full
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report all JVM diagnostics except per-thread stack traces
      tags:
      - Service diagnostics
  /insight/gc:
    get:
      operationId: gc
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report JVM garbage-collection counts
      tags:
      - Service diagnostics
  /insight/memory:
    get:
      operationId: memory
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report JVM memory use
      tags:
      - Service diagnostics
  /insight/system:
    get:
      operationId: system
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report operating-system information
      tags:
      - Service diagnostics
  /insight/thread-details:
    get:
      operationId: threadDetails
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report details and stack traces for every JVM thread
      tags:
      - Service diagnostics
  /insight/threads:
    get:
      operationId: threads
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: The requested JVM diagnostic report
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Forbidden
      security:
      - api_key: []
      summary: Report JVM thread counts
      tags:
      - Service diagnostics
components:
  schemas:
    CedarError:
      type: object
      additionalProperties: false
      description: A CEDAR error response. Diagnostic fields are populated only when
        they are relevant to the failure.
      properties:
        entities:
          type: object
          additionalProperties:
            type: object
            description: Named domain entities associated with exception-mapped errors.
            nullable: true
          description: Named domain entities associated with exception-mapped errors.
          nullable: true
        errorId:
          type: string
          format: uuid
          description: Correlation identifier for an internal error recorded in server
            logs.
          nullable: true
        errorKey:
          type: string
          description: Stable machine-readable error identifier.
          enum:
          - templateElementNotCreated
          - templateElementNotFound
          - templateElementNotDeleted
          - templateElementNotUpdated
          - templateElementsNotListed
          - templateFieldNotCreated
          - templateFieldNotFound
          - templateFieldNotDeleted
          - templateFieldNotUpdated
          - templateFieldsNotListed
          - templateNotCreated
          - templateNotFound
          - templateNotDeleted
          - templateNotUpdated
          - templatesNotListed
          - templateInstanceNotCreated
          - templateInstanceNotFound
          - templateInstanceNotDeleted
          - templateInstanceNotUpdated
          - templateInstancesNotListed
          - noReadAccessToResource
          - noWriteAccessToResource
          - noReadAccessToFolder
          - noWriteAccessToFolder
          - noReadAccessToTemplate
          - noWriteAccessToTemplate
          - noReadAccessToTemplateElement
          - noWriteAccessToTemplateElement
          - noReadAccessToTemplateField
          - noWriteAccessToTemplateField
          - noReadAccessToTemplateInstance
          - noWriteAccessToTemplateInstance
          - noReadAccessToArtifact
          - noWriteAccessToArtifact
          - noReadAccessToCategory
          - noWriteAccessToCategory
          - folderNotFound
          - artifactNotFound
          - artifactPreconditionRequired
          - artifactHasMovedOn
          - verbatimWriteRefused
          - nodeNotFound
          - unknownResourceType
          - missingParameter
          - invalidInput
          - sourceFolderNotFound
          - sourceResourceNotFound
          - targetFolderNotFound
          - nodeNotMoved
          - groupNotFound
          - userNotFound
          - uniqueConstraintCollision
          - invalidData
          - notAuthorized
          - permissionMissing
          - tokenInvalid
          - userInfoLoadByTokenFailed
          - userInfoLoadByApiKeyFailed
          - cedarUserNotFound
          - authorizationNotFound
          - permissionNotOwned
          - authorizationTypeUnknown
          - apiKeyNotFound
          - tokenMissing
          - tokenExpired
          - parentFolderNotSpecified
          - parentFolderSpecifiedTwice
          - pathNotNormalized
          - parentFolderNotFound
          - updateInvalidFolderName
          - createInvalidFolderName
          - nodeAlreadyPresent
          - folderNotCreated
          - missingNameAndDescription
          - missingData
          - folderNotDeleted
          - artifactNotDeleted
          - resourceNotCreated
          - invalidResourceType
          - invalidArtifactType
          - readOtherProfileForbidden
          - updateOtherProfileForbidden
          - folderCanNotBeDeleted
          - folderCanNotBeChanged
          - groupAlreadyPresent
          - groupCanBeModifiedOnlyByGroupAdmin
          - groupCanBeDeletedOnlyByGroupAdmin
          - groupMembersCanBeReadOnlyByGroupAdmin
          - groupRequiresAdministrator
          - groupUsersNotUpdated
          - specialGroupCanNotBeDeleted
          - folderPermissionsCanNotBeChanged
          - unknownInstanceOutputFormat
          - folderCopyNotAllowed
          - methodNotImplemented
          - upstreamServerError
          - nothingToDo
          - parentCategoryNotFound
          - categoryAlreadyPresent
          - categoryNotFound
          - categoryCanNotBeDeleted
          - rootCategoryCanNotBeDeleted
          - unableToAttachCategory
          - unableToDetachCategory
          - noCategoriesWereAttached
          - malformedJsonRequestBody
          - malformedSearchTerm
          - publishedArtifactCanNotBeChanged
          - publishedArtifactCanNotBeDeleted
          - versioningOnlyOnLatest
          - versioningOnlyByOwner
          - nonVersionedArtifactType
          - createDraftOnlyFromPublished
          - publishOnlyDraft
          - draftNotCreated
          - contentNotValid
          - doiNotSupportedByResourceType
          - doiCanNotBeSetForEmptyAtId
          - doiCanNotBeAltered
          - doiCanNotBeSet
          - doiAlreadyExists
          - dataCiteDOIDisabled
          - pinnedVersionUnavailable
          - resourceNotFound
          - internalError
          nullable: true
        errorReasonKey:
          type: string
          description: More specific machine-readable reason for the error.
          enum:
          - nonEmptyFolder
          - nonEmptyCategory
          - userHomeFolder
          - systemFolder
          - templateReferencedInInstances
          - validationError
          - continuationExpired
          nullable: true
        errorType:
          type: string
          description: Broad category of the error.
          enum:
          - notFound
          - invalidArgument
          - authorization
          - permission
          - server
          - validationError
          nullable: true
        message:
          type: string
          description: Human-readable error message.
          nullable: true
        objects:
          type: object
          additionalProperties:
            type: object
            description: Named structured values associated with the error.
          description: Named structured values associated with the error.
        operation:
          type: object
          description: Operation that failed. Its fields depend on the operation type.
          nullable: true
        parameters:
          type: object
          additionalProperties:
            type: object
            description: Named scalar values associated with the error.
          description: Named scalar values associated with the error.
        status:
          type: string
          description: Symbolic HTTP response status.
          enum:
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - METHOD_NOT_ALLOWED
          - NOT_ACCEPTABLE
          - CONFLICT
          - PRECONDITION_FAILED
          - UNSUPPORTED_MEDIA_TYPE
          - UNPROCESSABLE_ENTITY
          - PRECONDITION_REQUIRED
          - INTERNAL_SERVER_ERROR
          - NOT_IMPLEMENTED
          - BAD_GATEWAY
          - SERVICE_UNAVAILABLE
          - HTTP_VERSION_NOT_SUPPORTED
        statusCode:
          type: integer
          format: int32
          description: Numeric HTTP response status.
          maximum: 599
          minimum: 400
        suggestedAction:
          type: string
          description: Recovery action suggested to the client.
          enum:
          - none
          - requestRole
          - logout
          - logoutImmediately
          - provideAuthorizationHeader
          - refreshToken
          nullable: true
      required:
      - status
      - statusCode
    PopulatedField:
      type: object
      description: A field that has already been populated in the metadata.
      properties:
        fieldPath:
          type: string
          description: "Example: disease"
        fieldValueLabel:
          type: string
          description: "Example: atopic dermatitis"
        fieldValueType:
          type: string
          description: "Example: atopic dermatitis"
      required:
      - fieldPath
      - fieldValueLabel
    RecommendationInput:
      type: object
      additionalProperties: false
      description: The input used to request metadata recommendations for a target
        field.
      properties:
        includeDetails:
          type: boolean
          description: "If set to true, it returns details about how the recommendations\
            \ were generated."
        populatedFields:
          type: array
          description: The fields that have already been populated in the metadata.
          items:
            $ref: "#/components/schemas/PopulatedField"
        strictMatch:
          type: boolean
          description: "If set to true, only exact matches are considered."
        targetField:
          $ref: "#/components/schemas/TargetField"
        templateId:
          type: string
          description: "Example: https://repo.metadatacenter.org/templates/8bc64av5-df6b-48c8-8c61-6c016245918e"
    TargetField:
      type: object
      description: The field for which recommendations are requested.
      properties:
        fieldPath:
          type: string
          description: "Example: tissue"
  securitySchemes:
    api_key:
      description: "Your CEDAR API key, using the format 'apiKey <yourApiKey>'"
      in: header
      name: Authorization
      type: apiKey